requirements.txt
flask
torch
torchvision
Pillow
albumentations
numpy